fix: move common api functions to a separate file
