--- packages/netlify-cms-proxy-server/src/semaphore.d.ts
@@ -1,5 +1,0 @@
-declare module 'semaphore' {
-  export type Semaphore = { take: (f: Function) => void; leave: () => void };
-  const semaphore: (count: number) => Semaphore;
-  export default semaphore;
-}
